Move sharing link related methods into SharingLinkService
Creates circular dependencies, but these get compiled into one file on deploy so it shouldn't be a big issue - and the only circular dependencies are on static methods
